web/requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
Django==2.2.5
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-filters==0.2.1
django-guardian==2.1.0
django-redis==4.11.0
djangorestframework==3.10.3
drf-nested-routers==0.91
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
passlib==1.7.1
psycopg2==2.8.3
Pygments==2.4.2
pyquery==1.4.1
pytz==2019.2
redis==3.4.1
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
uWSGI==2.0.18
